paths & renames
