+++ fast64_internal/oot/c_writer/oot_room_writer/oot_object_to_c.py
